Created basic architecture for the API service layer
